Partial reverison back to UUIDs
Have changed ids back to UUIDs. This will allow me to return the
message_id _before_ execution when in batch mode as UUIDv4 can be
generated client-side.

Have also removed the "Deferred" utility class. Now return a reference
to an actual promise that can be directly awaited should the result of
the batched message-create be required.